% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.behavior.R
\name{customer.behavior}
\alias{customer.behavior}
\title{Shiny App for General Frequent Patterns}
\usage{
customer.behavior(Transaction)
}
\arguments{
\item{Data}{set (transaction data)}
}
\value{
Shiny App
}
\description{
Showing the Frequent Patterns with Shiny App.
}
\examples{
data("Groceries")
customer.behavior(Groceries)
}
\author{
Anh Hoang Duc (anhhd3@vpbank.com.vn)
}
\keyword{Association}
\keyword{Rules}
\keyword{Shiny,}

