% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.behavior.rooted.R
\name{customer.behavior.rooted}
\alias{customer.behavior.rooted}
\title{Shiny App for Frequent Patterns}
\usage{
customer.behavior.rooted(Transaction)
}
\arguments{
\item{the}{data set}
}
\value{
Shiny App
}
\description{
Showing the Frequent Patterns with Shiny App for 
rooted product
}
\examples{
data("Groceries")
customer.behavior.rooted(Groceries)
}
\author{
Anh Hoang Duc (anhhd3@vpbank.com.vn)
}
\keyword{Shiny}

