% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.convert.R
\name{transaction.convert}
\alias{transaction.convert}
\title{Quickly convert data frame to transaction}
\usage{
transaction.convert(data, type = c("wide", "long"))
}
\arguments{
\item{Data}{set - This functions supports 2 kinds of data wide & long}

\item{Type}{- Two option: "wide" and "long"}
}
\value{
Return a transaction data
}
\description{
This function helps users to convert data frame to
transaction data quickly
}
\examples{
df1 <- data.frame(Transaction_ID = c(1,2,1,3,1,2), Products = c(2,3,2,4,4,5))
transaction.convert(df1, type = "long")
df2 <- data.frame(ID = c(1,2,3), Product_1 = c(1,1,0), Product_2 = c(0,1,1), Product_3 = c(0,0,1))
transaction.convert(df2, type = "wide")
}
\author{
Anh Hoang Duc (anhhd3@vpbank.com.vn)
}
\keyword{Data}
\keyword{converter}

