% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPBank.R
\docType{package}
\name{VPBank-package}
\alias{VPBank}
\alias{VPBank-package}
\title{Customers' basket analysis with Shiny app}
\description{
VPBank package provides a convinient way to explore
customers' behavior using basket analysis with Shiny. 
This package is especially created for Data Analysts in 
Business Intelligence Competency Center (BICC) - VP Bank.
}
\author{
Anh Hoang Duc (anhhd3@vpbank.com.vn)
}

